refactored server to be fully multi threaded
